Move some model utils into a `common.py` module
